Attempting to fix nested bullets
When `introduction.txt` is rendered on https://leanpub.com/fullstacklisp/read, lines 16 through 18 appear as part of line 15 instead of on their own lines.  I've attempted to fix this, but note that I'm doing it blindly, I haven't rendered the content after my changes... So please do that before accepting this change.